+++ primitive/__main__.py
@@ -1,12 +1,13 @@
 import time
-from searches import search_with_for, search_with_findall, search_with_find
+
+from .searches import search_with_find, search_with_findall, search_with_for
 
 # Основная функция
 if __name__ == "__main__":
     while True:
-        folder_path = 'data'
+        folder_path = "data"
         input_text = input("Введите слово или несколько слов: ")
-        words = input_text.split('  ')  # Пробелы между словами считаются словосочетаниями
+        words = input_text.split("  ")  # Пробелы между словами считаются словосочетаниями
 
         # Запуск поиска с использованием разных методов
         start_time = time.time()
@@ -15,7 +16,6 @@ if __name__ == "__main__":
         print("Результаты для цикла for:")
         print(result_for)
         print(f"Выполнено за {(end_time - start_time)*1000} мс\n")
-
 
         start_time = time.time()
         result_findall = search_with_findall(words, folder_path)
@@ -30,4 +30,3 @@ if __name__ == "__main__":
         print("Результаты для метода find:")
         print(result_find)
         print(f"Выполнено за {(end_time - start_time)*1000} мс\n")
-
